nD simultaneous extremes probs
- Test simultaneous extremes for all possible combinations in a given
stn_comb. Starting from 2 to len(stn_comb).
- A new function in _fill_simult_freqs_arr computes this. It is
generalization of what _fill_freqs_arr does. Should be faster than the
older one.
- The number of possible combinations can sky rocket pretty fast. So be
careful with the maximum number of stations.
- Values for all the simulations are written to the hdf5. This allows
for the visualization of the CDF of these probabilities.
- Plotting has been changed a bit. Previous cluster plotting functions
are called binary and the new ones are called nD. The nD plotting
functions have yet to be parallelized.